feat: standardize qr code for keysign
