+++ frontend/src/vault/keysign/start/peerDiscovery/KeysignPeerDiscoveryStep/KeysignPeerDiscoveryStep.styled.tsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const Content = styled.div`
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 40px;
+`;
+
+export const Wrapper = styled.div`
+  flex: 1;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 40px;
+`;
